refactor(User): JdbcTemplate -> spring data jap 적용
