+++ src/main/java/gift/domain/user/repository/JpaUserRepository.java
@@ -1,0 +1,14 @@
+package gift.domain.user.repository;
+
+import gift.domain.user.User;
+import gift.domain.user.dto.UserDTO;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface JpaUserRepository extends JpaRepository<User, Long> {
+
+    boolean existsByEmail(String email);
+
+    User findByEmailAndPassword(String email, String password);
+}
