+++ src/main/java/gift/domain/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package gift.domain.user;
+package gift.domain.user.repository;
 
+import gift.domain.user.User;
 import gift.domain.user.dto.UserDTO;
 
 public interface UserRepository {
