+++ src/main/resources/application.properties
@@ -13,3 +13,10 @@ logging.level.org.springframework.jdbc.core.StatementCreatorUtils=TRACE
 spring.sql.init.mode=always
 spring.sql.init.schema-locations=classpath:schema.sql
 spring.sql.init.data-locations=classpath:data.sql
+
+# spring data jpa
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.show-sql=true
+
+# hibernate
+spring.jpa.hibernate.ddl-auto=create
